requirements.txt
kafka-python
python-psycopg2
requests
jsonpickle
python-dotenv